// src/enums/game.enum.ts
export enum GameDifficultyEnum {
    EASY = 'easy',
    MEDIUM = 'medium',
    HARD = 'hard',
}
export enum GameCategoryEnum {
    PUZZLE = 'puzzle',
    ACTION = 'action',
    ARCADE = 'arcade',
    STRATEGY = 'strategy',
}
export enum GameLeaderboardPrizeTypeEnum {
    POINT = 'POINT',
    TOKEN = 'TOKEN',
}
export enum PrizeTypeEnum {
    POINT = 'POINT',
    TOKEN = 'TOKEN',
    NONE = 'NONE',
    EXTRA_SPIN = 'EXTRA_SPIN',
    DOUBLE_POINTS_24H = 'DOUBLE_POINTS_24H',
  }
  
  export enum LeaderboardFilterTypeEnum {
    Daily = 'daily',
    AllTime = 'all_time',
    //...
  }